Remove special-cased darwin compiler-rt support.
We ship a set of bionic headers with our clang, even on darwin. As a
result, these clang implementations of atomics are not compatible with
stdatomic.h. On a normal darwin machine, the compiler needs to provide
these. For one running our modified clang, it may not do so.

This patch is potentially brittle, as any update to compiler-builtins
will cause the hash-deletion portion of this patch to fail to apply.

This patch does not belong upstream, as it is an oddity of Android's
toolchain.

Bug: 146785396
Change-Id: Ic3b19382b5796ac93e829e55c375fdaf4dd2ed57