Trim trailing whitespace from command output to fix mac builds.
We were getting lucky in the past that the result of running our SDK
sysroot detection was the last thing on the command line. Unfortunately,
we now append further arguments, so this splits our compiler wrapper
incorrectly into 2 separate commands. Trimming the trailing whitespace
from the command output fixes the problem.

Test: DIST_DIR=dist toolchain/android_rust/build.py
Change-Id: Ief9f49b95385ca5302d919e37145434b59f47fe5