Add libc++ search path for Android checkouts
Previously, we only set the rpath to the absolute path used during the
build. Relativizing this would be insufficient, as rustc is run from a
different location while compiling itself than it will live in the final
Android tree. In this patch, I add a relative search path to help rustc
find libc++ in the Android tree.

Bug: 165026104
Change-Id: I41b86d6e2a0ceb95a925afaa7d149e148a9933a3