cargo: Don't list-files in bare git repo
This change was mistakenly removed in aosp/1403447 because
it didn't apply cleanly and the bug appeared to have been fixed
by an upstream change. Upon closer inspection, the bug still exists
so this change needs to be reapplied with the modification of using
Ok() to return None to match the refactored code.

Bug: 165026104
Test: TH along with 1.45.2 upgrade
Change-Id: I7b5abd81de80c936f73655d12fb090cd453597d2