Silence warnings about unused arguments
This CL silences thousands of warnings about unused `-fuse-ld=lld`
arguments.  The granularity of configuration means that we can't pass
this argument only to the linker step.  Instead we have to add it to the
compiler wrapper functions, causing it to be present when the compiler
is invoked to generate object files.

If a way is found to only pass this argument to linker steps the warning
should be re-enabled.

Test: ./do_build.py
Change-Id: Ie32f2729d453f01b3464a914651eca7c8c013de3