Add patch to disable recording of function records in coverage maps
Bug: http://b/186873963

Function addresses are needed for value profiling of indirect calls.
Value profiling is disabled by default and not planned to be used in
Android.  However, recording function addresses prevents them from being
stripped.  This CL adds a local patch to disable recording function
addresses when value profiling is not used.  A similar patch exists for
the clang toolchain (aosp/1235306) as well.

(There's work in upstream to allow linker to discard functions
even if their addresses are recorded in coverage metadata but that
hasn't been validated for Android yet.)

Test: Build aosp/1768548 with code coverage instrumentation and verify
      that lib_ld_discard_test.rs:fn2 is discarded by the linker.
Change-Id: Id576a214bff3e5ff4b7fb093bfbd25dd9b2c42d4